constraints middleware
The constraints middleware would allow of different
constraints to be applied to different storage policies.
In the case of swiftonfile this is important, because
the constraints are different than the existing constraints
of the default swift constraints. For example: Sof needs
to check for the size of each directory in a object path,
which is not necessary for default swift.

Signed-off-by: Thiago da Silva <thiago@redhat.com>