Add check for content-type header in request
the default check_object_creation checks if content-type
header is in the request headers. If it is not a HTTPBadRequest
returns. The proxy has code to detect and add the content-type,
so for now we have moved this code here, so that the constraints
middleware can be in the proxy pipeline.

Signed-off-by: Thiago da Silva <thiago@redhat.com>