new swiftonfile only constraints middleware
Instead of having a generic constraints middleware where
admins would have to configure what functions to call in the
configuration file, we are proposing a new approach where
the middleware is specific to a storage policy type.
In the case of swiftonfile we currently just add some additional
checks on the object name.

This middleware could be further expanded to include checks on
container creation too, but similar changes would need to be
added to controlloer/container code in upstream swift.

Signed-off-by: Thiago da Silva <thiago@redhat.com>